Cast to argument type to avoid unexpected behavior

+++ src/main/java/org/magcode/sem6000/connector/receive/DataDayResponse.java
@@ -15,7 +15,7 @@ public class DataDayResponse extends SemResponse {
     int today = 0;
     for (int i = 0; i < 24; i = i + 1) {
       byte[] cv = new byte[]{values[i * 2], values[(i * 2) + 1]};
-      LocalDateTime timeOfValue = LocalDateTime.now().plusHours(-23 + i).withMinute(0).withSecond(0).withNano(0);
+      LocalDateTime timeOfValue = LocalDateTime.now().plusHours(-23 + (long) i).withMinute(0).withSecond(0).withNano(0);
 
       int hourVal = new BigInteger(cv).intValue();
       if (now.getDayOfMonth() == timeOfValue.getDayOfMonth()) {
@@ -27,16 +27,16 @@ public class DataDayResponse extends SemResponse {
     this.enToday = today;
   }
 
-	public String toString() {
+  public String toString() {
     return "DataDay Energy last 24h: " + en24h + " Wh, Energy today: " + enToday + " Wh";
-	}
+  }
 
-	public int getLast24h() {
-		return en24h;
-	}
+  public int getLast24h() {
+    return en24h;
+  }
 
-	public int getToday() {
-		return enToday;
-	}
+  public int getToday() {
+    return enToday;
+  }
 
 }
